FE-1239 - default type in create domain form
