+++ src/pages/domains/components/CreateForm.js
@@ -21,11 +21,7 @@ import useDomains from '../hooks/useDomains';
 import { DomainAlignmentModal } from './DomainAlignmentModal';
 import useModal from 'src/hooks/useModal';
 
-const initFilters = {
-  type: { defaultValue: 'sending' },
-};
-
-export default function CreateForm() {
+export default function CreateForm({ defaultType = 'sending' }) {
   const history = useHistory();
   const { filters } = usePageFilters(initFilters);
   const { type } = filters;
@@ -51,7 +47,7 @@ export default function CreateForm() {
 
   const { control, register, handleSubmit, errors, watch, setValue } = useForm({
     defaultValues: {
-      primaryUse: getDomainCreateType(type),
+      primaryUse: defaultType,
       assignTo: 'shared',
     },
   });
