+++ src/pages/domains/components/SendingDomainsEmptyState.js
@@ -1,30 +1,21 @@
 import React from 'react';
-import { PageLink } from 'src/components/links';
 import { EmptyState } from 'src/components/matchbox';
-import SendingMailWebp from '@sparkpost/matchbox-media/images/Sending-Mail.webp';
-import SendingMail from '@sparkpost/matchbox-media/images/Sending-Mail@medium.jpg';
-import { Box } from 'src/components/matchbox';
-import { Bold } from 'src/components/text';
+import ConfigurationWebp from '@sparkpost/matchbox-media/images/Configuration.webp';
 import { TranslatableText } from 'src/components/text';
+// import { ExternalLink, PageLink } from 'src/components/links';
+// import { useHistory } from 'react-router-dom';
 import { LINKS } from 'src/constants';
 
 export default function SendingDomainsEmptyState() {
+  // const history = useHistory();
   return (
     <EmptyState>
       <EmptyState.Header>Sending Domains</EmptyState.Header>
       <EmptyState.Content>
-        <Box mb="400">
-          <p>
-            Sending domains are used to indicate who an email is from via the "From" header. DNS
-            records can be configured for a sending domain, which allows recipient mail servers to
-            authenticate messages sent from SparkPost.
-          </p>
-        </Box>
         <p>
-          <Bold>
-            At least one verified sending domain is required in order to start sending or enable
-            analytics.
-          </Bold>
+          Sending domains are used to indicate who an email is from via the "Form" records can be
+          configured for a sending domain, which allows recipient mail servers to authenticate
+          messages sent from SparkPost.
         </p>
         <EmptyState.List>
           <li>
@@ -42,11 +33,11 @@ export default function SendingDomainsEmptyState() {
           </li>
         </EmptyState.List>
       </EmptyState.Content>
-      <EmptyState.Image src={SendingMail} seeThrough>
-        <source srcSet={SendingMailWebp} type="image/webp" />
-      </EmptyState.Image>
-      <EmptyState.Action component={PageLink} to="/domains/create?type=sending">Add Sending Domain</EmptyState.Action>
-      <EmptyState.Action variant="outline" to={LINKS.SENDING_DOMAIN_DOCS} external>
+      <EmptyState.Image src={ConfigurationWebp} />
+      <EmptyState.Action to="/domains/create">Add Sending Domain</EmptyState.Action>
+
+      {/* TODO: FIX TO LINK */}
+      <EmptyState.Action variant="outline" to={LINKS.AB_TESTING_DOCS} external>
         Sending Domains Documentation
       </EmptyState.Action>
     </EmptyState>
